fit frmtaikhoan, frmnv, fit bug basic

+++ WindowsFormsApp9/frmquanlytaikhoan.cs
@@ -35,27 +35,56 @@ namespace WindowsFormsApp9
         {
             using (DataClasses1DataContext dt = new DataClasses1DataContext())
             {
-                TaiKhoan themTK = new TaiKhoan();
-                themTK.HovaTen = tb_HoTen.Text;
-                themTK.ChucVu = tb_ChucVu.Text;
-                themTK.TenTaiKhoan = tb_User.Text;
-                themTK.MatKhau = tb_Pass.Text;
-                dt.TaiKhoans.InsertOnSubmit(themTK);
-                dt.SubmitChanges();
-            }
-            loadDataTaiKhoan();
+                if(tb_ChucVu.Text == "" || tb_HoTen.Text == "" || tb_Pass.Text == "" || tb_User.Text == "")
+                {
+                    MessageBox.Show("Bạn chưa nhập đủ thông tin");
+                }    
+                try 
+                {
+                    TaiKhoan themTK = new TaiKhoan();
+                    themTK.HovaTen = tb_HoTen.Text;
+                    themTK.ChucVu = tb_ChucVu.Text;
+                    themTK.TenTaiKhoan = tb_User.Text;
+                    themTK.MatKhau = tb_Pass.Text;
+                    dt.TaiKhoans.InsertOnSubmit(themTK);
+                    dt.SubmitChanges();
+                    loadDataTaiKhoan();
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.Message);
+                }
+            }          
         }
 
         private void btn_Xoa(object sender, EventArgs e)
         {
             using (DataClasses1DataContext dt = new DataClasses1DataContext())
             {
-                string id = dataView.SelectedCells[0].OwningRow.Cells["TenTaiKhoan"].Value.ToString();
-                TaiKhoan delete = dt.TaiKhoans.Where(p => p.TenTaiKhoan.Equals(id)).SingleOrDefault();
-                dt.TaiKhoans.DeleteOnSubmit(delete);
-                dt.SubmitChanges();
+                if (tb_User.Text == "")
+                {
+                    MessageBox.Show("Bạn chưa nhập đủ thông tin");
+                    return;
+                }                   
+                try 
+                {
+                    TaiKhoan delete = dt.TaiKhoans.FirstOrDefault(a => a.TenTaiKhoan == tb_User.Text);
+                    if(delete != null)
+                    {
+                        dt.TaiKhoans.DeleteOnSubmit(delete);
+                        dt.SubmitChanges();
+                        loadDataTaiKhoan();
+                    }
+                    else
+                    {
+                        MessageBox.Show("Không tồn tại tài khoản này");
+                    }                        
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.Message);
+                }
             }
-            loadDataTaiKhoan();
         }
 
         private void btn_Khoitao(object sender, EventArgs e)
@@ -66,9 +95,12 @@ namespace WindowsFormsApp9
             tb_Pass.Text = "";
         }
 
-        private void dataView_CellContentClick(object sender, DataGridViewCellEventArgs e)
+        private void dataView_Click(object sender, EventArgs e)
         {
-
+            tb_HoTen.Text = dataView.CurrentRow.Cells["HovaTen"].Value.ToString();
+            tb_ChucVu.Text = dataView.CurrentRow.Cells["ChucVu"].Value.ToString();
+            tb_User.Text = dataView.CurrentRow.Cells["TenTaiKhoan"].Value.ToString();
+            tb_Pass.Text = dataView.CurrentRow.Cells["MatKhau"].Value.ToString();
         }
     }
 }
